feat: (Day) Player variables can now be configured to be recordable in a player's records

+++ Intersect.Server/Entities/Npc.cs
@@ -207,7 +207,7 @@ namespace Intersect.Server.Entities
 
                 if (killer is Player playerKiller)
                 {
-                    int recordKilled = playerKiller.IncrementRecord(RecordType.NpcKilled, Base.Id);
+                    long recordKilled = playerKiller.IncrementRecord(RecordType.NpcKilled, Base.Id);
                     if (Options.SendNpcRecordUpdates && recordKilled % Options.NpcRecordUpdateInterval == 0)
                     {
                         playerKiller.SendRecordUpdate(Strings.Records.enemykilled.ToString(recordKilled, Name));
